requirements.txt
flask
gunicorn
tensorflow==2.12.0
pillow
numpy